Extract navigation toggle suppression policy
